requirements.txt
django>=4.2
djangorestframework>=3.15.0
pyyaml
requests
django-cors-headers